Backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
sqlalchemy==2.0.25
python-dotenv==1.0.0
groq==0.4.3
python-jose==3.3.0
aiofiles==23.2.1
python-socketio==5.10.0
